commit changes

+++ src/main/java/ru/jpa/tests/TestOneToManyPerson/test/HibernateTestPersonOneToMany.java
@@ -2,14 +2,14 @@ package ru.jpa.tests.TestOneToManyPerson.test;
 
 import ru.jpa.tests.TestOneToManyPerson.dao.PersonOneToManyDaoHibernate;
 import ru.jpa.tests.TestOneToManyPerson.model.AddressOneToMany;
-import ru.jpa.tests.TestOneToManyPerson.model.PersonOneToMany;
+import ru.jpa.tests.TestOneToManyPerson.model.PersonOneToManyHib;
 import ru.jpa.tests.imp.ITestOneToManyAndCollection;
 
 import java.util.List;
 
 public class HibernateTestPersonOneToMany implements ITestOneToManyAndCollection {
     private PersonOneToManyDaoHibernate dao;
-    private List<PersonOneToMany> personsWithAddress;
+    private List<PersonOneToManyHib> personsWithAddress;
     private AddressOneToMany address;
 
     @Override
@@ -19,18 +19,18 @@ public class HibernateTestPersonOneToMany implements ITestOneToManyAndCollection
     }
 
     @Override
-    public List<PersonOneToMany> getAllPersons() {
+    public List<PersonOneToManyHib> getAllPersons() {
         dao = new PersonOneToManyDaoHibernate();
-        List<PersonOneToMany> persons = dao.getAllPersons();
+        List<PersonOneToManyHib> persons = dao.getAllPersons();
         persons.forEach(person -> System.out.println(person));
         return persons;
     }
 
-    public List<PersonOneToMany> getAllPersonWithAddress() {
+    public List<PersonOneToManyHib> getAllPersonWithAddress() {
         dao = new PersonOneToManyDaoHibernate();
         personsWithAddress = dao.getAllPersons();
 
-        for (PersonOneToMany person: personsWithAddress) {
+        for (PersonOneToManyHib person: personsWithAddress) {
             System.out.println(person);
 //            System.out.println(person.getAddresses());
         }
@@ -40,14 +40,16 @@ public class HibernateTestPersonOneToMany implements ITestOneToManyAndCollection
     @Override
     public void deletePerson() {
         dao = new PersonOneToManyDaoHibernate();
-        List<PersonOneToMany> persons = dao.getAllPersons();
-        persons.forEach(person -> dao.delete(person));
+        List<PersonOneToManyHib> persons = dao.getAllPersons();
+        if(!persons.isEmpty() && persons != null){
+            persons.forEach(person -> dao.delete(person));
+        }
     }
 
     public void updatePerson() {
         dao = new PersonOneToManyDaoHibernate();
-        List<PersonOneToMany> persons = dao.getAllPersons();
-        for (PersonOneToMany person : persons) {
+        List<PersonOneToManyHib> persons = dao.getAllPersons();
+        for (PersonOneToManyHib person : persons) {
             person.setFirstName("testName");
             dao.update(person);
         }
